bitpocket can work without a remote shell
This enables bitpocket to `init` and acquire and release the remote lock
without using a remote shell. This will bitpocket to operate in areas
where a remote shell is not allowed or not practical (for instance, if
the remote is an embedded or non-posix system).

It also attempts to avoid a race condition between the time when a
remote lock is requested and created by placing locks for different
clients in different files. This will allow other potential sync targets
to notice a requested lock file while the local client is inspecting
other lock files to determine if it is safe to continue with the sync.