sync: Add expiration date to remote lock files
This allows remote lock files to automatically expire after a certain
period of time. That amount of time can be configured per client and as
implemented here defaults indefinite (ie. not expire at all), but can be
changed in the config to a more reasonable value such as 30 minutes.
After this time, the lock file will automatically expire by all other
clients and syncs will be allowed.